Change how policies are applied
Instead of adding the policy type in the endpoint interface you now add a separate Enforce<> interface. I also removed the requirement to inject the HttpContext in the constructor. This is a convenience when writing endpoint classes. Now you don't have to specify a constructor and pass down to the base class. The dependency injection framework will handle injecting the necessary properties.