Create HandlerFactory
Instead of trying to resolve all the dependencies in Voyager, I created a HandlerFactory. If anyone wants to create an instance of a handler they can use the factory which will handle wiring up all the properties correctly. MediatR resolves from the interface which will use the factory as well.