Inject HttpContext, Swashbuckle, ValueProviders
If a mediator handler implements the IInjectHttpContext a proxy will be created so that the HttpContext can be auto injected.
Added a Voyager.Swashbuckle package to make using Swagger better and easier.
Removed all of the SetProperties classes and now just use Value Providers with JsonDeserialization to bind to request objects.